Third-provider base: substrate (P0 lint + P1 VFS + P2 clock/RNG + P4 fd
fixtures/guard/dataplane) + the transport seam (pqcomm_stdio provider,
Transport selection, --stdio-wire, wire_session_initialize).

Conflict resolutions (composed, not side-picked where both mattered):
- pgrcolumnar part_cache/reader: substrate's vfs::stat FileInfo identity
  wins; the wasm MetadataExt cfg imports are obsolete (fields, not methods).
- pgrcolumnar segfile: substrate's SegFd/vfs rewiring wins; the p5 wasm
  map_segs twin is re-expressed over vfs::pread on SegFd (EINTR loop
  matching SegFile::read); posix mmap arm keeps the P1 carve-out comment,
  now cfg(not(wasm)).
- fd sync.rs: p5 wasm no-op pg_flush_data arm + substrate carve-out comment
  on the (not(linux), not(wasm)) msync arm.
- fd copydir.rs: substrate's vfs-routed mkdir_p wins (wasm-compatible).
- transam_xlog write.rs: composed — p5's init_small process_id seam (wasm
  has no pids) + substrate's fd::pg_unlink (vfs-routed).
- basebackup: substrate's vfs::FileInfo portable fields win.
- timeout: both comments + p5's cfg(not(wasm)) on the timer spawn.
- tcop/postgres Cargo.toml: dropped merge-duplicated timestamp_seams dep.
- notes/LANES.md: kept both sides' rows.

Determinism-lint reconciliation (allowlist == census at this tip, PASS 0/0):
- NEW annotated: pqcomm_stdio fs 2 (the seam's declared provider row),
  waiteventset wasm_stub blocking 1 (wasm-only WES backend).
- Budgets grown w/ DST-REVIEW(p4-simnet) markers: tablespace fs 16->17,
  genfile builtins fs 7->10, lockfile fs 16->17, main_main env 1->2 (all =
  wasm/p5-boot cfg arms meeting the lint lineage for the first time).
- Budget shrunk per ratchet law: pqcomm socket.rs fs 14->13.

Both-cfg workspace check GREEN (default + --cfg pgrust_sim).

+++ crates/backend/executor/execgrouping/src/lib.rs
@@ -439,7 +439,9 @@ impl SimpleHashIndex {
     /// SH_START_ITERATE + SH_ITERATE: backward walk from the first free
     /// bucket. `cursor` starts at 0; it packs the start bucket (high 32
     /// bits) and visited count + 1 (low 32 bits).
-    fn iterate(&self, cursor: &mut usize) -> Option<u32> {
+    // Cursor packing is u64 (not usize: the high-32 shift truncates on
+    // 32-bit wasm); identical words on 64-bit targets.
+    fn iterate(&self, cursor: &mut u64) -> Option<u32> {
         let size = self.size() as usize;
         let (start, mut visited) = if *cursor == 0 {
             let start =
@@ -447,7 +449,7 @@ impl SimpleHashIndex {
                     as u32;
             (start, 0usize)
         } else {
-            ((*cursor >> 32) as u32, (*cursor & 0xffff_ffff) - 1)
+            ((*cursor >> 32) as u32, ((*cursor & 0xffff_ffff) as usize) - 1)
         };
         let mut result = None;
         while visited < size {
@@ -459,7 +461,7 @@ impl SimpleHashIndex {
                 break;
             }
         }
-        *cursor = ((start as usize) << 32) | (visited + 1);
+        *cursor = ((start as u64) << 32) | (visited as u64 + 1);
         result
     }
 
@@ -871,7 +873,7 @@ impl<'mcx> TupleHashTable<'mcx> {
     }
 
     /// C SH_START_ITERATE/SH_ITERATE bucket-order drain; `cursor` starts 0.
-    pub fn iterate(&self, cursor: &mut usize) -> Option<u32> {
+    pub fn iterate(&self, cursor: &mut u64) -> Option<u32> {
         self.hashtab.iterate(cursor)
     }
